переписал тестирование на тест метода вьюхи
